Restructure table components
